% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{params_series}
\alias{params_series}
\title{Inherit Documentation for time series distances package}
\arguments{
\item{X}{Numeric vector representing first time series}

\item{Y}{Numeric vector representing second time series}
}
\description{
Compute Time Warped Edit distance, Minimum Jump Cost or Fast Dynamitc Time Warping distances
}
