% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{PrunedDTWDistance}
\alias{PrunedDTWDistance}
\title{Pruned Dynamic Time Warping (Pruned DTW) Distance}
\usage{
PrunedDTWDistance(X, Y, window = 1)
}
\arguments{
\item{X}{Numeric vector representing first time series}

\item{Y}{Numeric vector representing second time series}

\item{window}{Window size}
}
\value{
Numerical distance between time series \code{X} and \code{Y}
}
\description{
Implements Dynamic Time Warping with Pruned Warping  Paths (PrunedDTW).The main speed-up comes from recognition and pruning cells 
in the DTW matrix that are guaranteed to not lead to alignments that will result in the optimal path.. 
Similar to Dynamic Time Warping (DTW), the distance between multiple time series can be calculated 
using \code{\link[=proxy::dist()]{proxy::dist()}}function in package \pkg{proxy} with the method given, 
or invoked on all pairs `x[i],y[j]` to build the distance matrix.
}
\examples{

a <- runif(100) *100
b <- a + 5
PrunedDTWDistance(a, b)

}
\references{
Silva D.F, Batista G. (2016) Speeding Up All-Pairwise Dynamic Time Warping Matrix Calculation. SDM 2016: 837-845
}
