% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{TWEDistance}
\alias{TWEDistance}
\title{Time Warped Edit Distance (TWED)}
\usage{
TWEDistance(X, Y, Xts=NULL, Yts = NULL, nu = 0.001, lambda = 1.0, 
                   degree = 2)
}
\arguments{
\item{X}{Numeric vector representing first time series}

\item{Xts}{Vector containing the time stamps for time series \code{X}. The length of this vector should match length of \code{X}}

\item{Y}{Numeric vector representing second time series}

\item{Yts}{Vector containing the time stamps for time series \code{Y}. The length of this vector should match length of \code{Y}}

\item{nu}{Elasticity parameter value, required to be 0 or more}

\item{lambda}{Penalty for deletion operation}

\item{degree}{Power degree for the evaluation of the local distance between samples; \code{degree > 0} is required}
}
\value{
Numerical distance between time series \code{X} and \code{Y}
}
\description{
Original implementation of Time Warped Edit Distance (TWED) for discrete time series \code{X} and \code{Y} matching, first proposed in 2009 by P.-F. Marteau. 
The function computes TWED as the optimal alignment between two time series \code{X} and \code{Y}, given as numeric
vectors. Lengths of \code{X} and \code{Y} may differ.
Similar to Dynamic Time Warping (DTW), the distance between multiple time series can be calculated 
using \code{\link[=proxy::dist()]{proxy::dist()}}function in package \pkg{proxy} with the method given, 
or invoked on all pairs `x[i],y[j]` to build the distance matrix.
}
\examples{

a <- runif(100) *100
b <- a + 5
TWEDistance(a, b)

}
\references{
Marteau P. F. (2009). "Time Warp Edit Distance with Stiffness Adjustment for Time Series Matching". IEEE Transactions on Pattern Analysis and Machine Intelligence. 31 (2): 306–318. arXiv:cs/0703033. doi:10.1109/TPAMI.2008.76.
}
